# docs/swagger.yaml
basePath: /v1
definitions:
  server.DNSRequest:
    properties:
      vel:
        example: "48.1"
        type: string
      x:
        example: "32.1"
        type: string
      "y":
        example: "4.2"
        type: string
      z:
        example: "31.3"
        type: string
    type: object
  server.DNSResponse:
    properties:
      loc:
        example: 27372.229
        type: number
    type: object
  server.ErrorResponse:
    properties:
      errors:
        example: reference not found
        type: string
      message:
        example: Not Found
        type: string
      status:
        example: 404
        type: integer
    type: object
host: localhost:9000
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://example.com/support
  description: galaxy offers a drone navigation service (DNS).
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://example.com/terms/
  title: galaxy API
  version: "1.0"
paths:
  /:
    get:
      description: Healthcheck endpoint
      operationId: home-galaxy
      produces:
      - text/plain
      responses:
        "200":
          description: Documentation message
          schema:
            type: string
      summary: Home
      tags:
      - galaxy
  /sectors/{sector_id}/dns:
    post:
      consumes:
      - application/json
      description: Get a DNS location of a sector of the galaxy
      operationId: dns-sector-galaxy
      parameters:
      - description: Sector ID
        in: path
        name: sector_id
        required: true
        type: integer
      - description: x, y, z are the coordinates and vel is the velocity
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/server.DNSRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.DNSResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
            type: object
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/server.ErrorResponse'
            type: object
        "406":
          description: Not Acceptable
          schema:
            $ref: '#/definitions/server.ErrorResponse'
            type: object
      summary: Show a DNS location
      tags:
      - galaxy
swagger: "2.0"